Move the AbstractHsmManager to the mock package
